Split off the .o5 stuff for easier testing/debugging.
