+++ tests/test_oims.py
@@ -133,166 +133,6 @@ class oims_tests(unittest.TestCase):
         db0.close()
         db1.close()
         
-    def test_o5_write(self):
-        """Test saving data to the OrvilleImageHDF5 format."""
-        
-        # Setup the file names
-        testFile = os.path.join(self.testPath, 'test.o5')
-        
-        db = OrvilleImageDB(oimsFile, 'r')
-        nf = OrvilleImageHDF5(testFile, 'w', imager_version=db.header.imager_version, 
-                              station=db.header.station)
-                                            
-        # Fill it
-        for rec in db:
-            hdr, img = rec
-            for key in hdr:
-                if isinstance(hdr[key], bytes):
-                    hdr[key] = hdr[key].decode()
-            nf.add_image(hdr, img)
-            
-        # Done
-        db.close()
-        nf.close()
-        
-        # Re-open
-        db0 = OrvilleImageDB(oimsFile, 'r')
-        db1 = OrvilleImageHDF5(testFile, 'r')
-        
-        # Validate
-        ## File header
-        for attr in ('imager_version', 'station', 'stokes_params', 'ngrid', 'nchan', 'flags'):
-            attr0 = getattr(db0.header, attr, None)
-            if isinstance(attr0, bytes):
-                attr0 = attr0.decode()
-            self.assertEqual(attr0, getattr(db1.header, attr, None))
-        for attr in ('pixel_size', 'start_time', 'stop_time'):
-            self.assertAlmostEqual(getattr(db0.header, attr, None), getattr(db1.header, attr, None), 6)
-        ## First image
-        ### Image header
-        hdr0, img0 = db0.read_image()
-        hdr1, img1 = db1.read_image(0)
-        for attr in ('stokes_params', 'ngrid', 'pixel_size', 'ngrid'):
-            try:
-                attr0 = hdr0[attr]
-                if isinstance(attr0, bytes):
-                    attr0 = attr0.decode()
-            except KeyError:
-                attr0 = None
-            self.assertEqual(attr0, getattr(hdr1, attr, None))
-        for attr in ('start_time', 'int_len', 'lst', 'start_freq', 'stop_freq', 'bandwidth', 'fill', 'center_ra', 'center_dec'):
-            try:
-                attr0 = hdr0[attr]
-                if isinstance(attr0, bytes):
-                    attr0 = attr0.decode()
-            except KeyError:
-                attr0 = None
-            self.assertAlmostEqual(attr0, getattr(hdr1, attr, None), 6)
-        ### Image
-        for i in range(img0.shape[0]):
-            for j in range(img0.shape[1]):
-                for k in range(img0.shape[2]):
-                    for l in range(img0.shape[3]):
-                        self.assertAlmostEqual(img0[i,j,k,l], img1[i,j,k,l], 6)                
-        ## First image another way
-        hdr1, img1 = db1[0]
-        for attr in ('stokes_params', 'ngrid', 'pixel_size', 'ngrid'):
-            try:
-                attr0 = hdr0[attr]
-                if isinstance(attr0, bytes):
-                    attr0 = attr0.decode()
-            except KeyError:
-                attr0 = None
-            self.assertEqual(attr0, getattr(hdr1, attr, None))
-        for attr in ('start_time', 'int_len', 'lst', 'start_freq', 'stop_freq', 'bandwidth', 'fill', 'center_ra', 'center_dec'):
-            try:
-                attr0 = hdr0[attr]
-                if isinstance(attr0, bytes):
-                    attr0 = attr0.decode()
-            except KeyError:
-                attr0 = None
-            self.assertAlmostEqual(attr0, getattr(hdr1, attr, None), 6)
-        ### Image
-        for i in range(img0.shape[0]):
-            for j in range(img0.shape[1]):
-                for k in range(img0.shape[2]):
-                    for l in range(img0.shape[3]):
-                        self.assertAlmostEqual(img0[i,j,k,l], img1[i,j,k,l], 6)
-                        
-        db0.close()
-        db1.close()
-        
-    def test_o5_compressed_write(self):
-        """Test saving data to the OrvilleImageHDF5 format with compression."""
-        
-        # Setup the file names
-        testFile = os.path.join(self.testPath, 'test.o5')
-        
-        db = OrvilleImageDB(oimsFile, 'r')
-        nf = OrvilleImageHDF5(testFile, 'w', imager_version=db.header.imager_version, 
-                              station=db.header.station, compression='gzip')
-                                            
-        # Fill it
-        for rec in db:
-            hdr, img = rec
-            for key in hdr:
-                if isinstance(hdr[key], bytes):
-                    hdr[key] = hdr[key].decode()
-            nf.add_image(hdr, img)
-            
-        # Done
-        db.close()
-        nf.close()
-        
-        # Re-open
-        db0 = OrvilleImageDB(oimsFile, 'r')
-        db1 = OrvilleImageHDF5(testFile, 'r')
-        
-        # Validate
-        ## File header
-        for attr in ('imager_version', 'station', 'stokes_params', 'ngrid', 'nchan', 'flags'):
-            attr0 = getattr(db0.header, attr, None)
-            if isinstance(attr0, bytes):
-                attr0 = attr0.decode()
-            self.assertEqual(attr0, getattr(db1.header, attr, None))
-        for attr in ('pixel_size', 'start_time', 'stop_time'):
-            self.assertAlmostEqual(getattr(db0.header, attr, None), getattr(db1.header, attr, None), 6)
-        ## First image
-        ### Image header
-        hdr0, img0 = db0.read_image()
-        hdr1, img1 = db1.read_image(0)
-        for attr in ('stokes_params', 'ngrid', 'pixel_size', 'ngrid'):
-            attr0 = getattr(hdr0, attr, None)
-            if isinstance(attr0, bytes):
-                attr0 = attr0.decode()
-            self.assertEqual(attr0, getattr(hdr1, attr, None))
-        for attr in ('start_time', 'int_len', 'lst', 'start_freq', 'stop_freq', 'bandwidth', 'fill', 'center_ra', 'center_dec'):
-            self.assertAlmostEqual(getattr(hdr0, attr, None), getattr(hdr1, attr, None), 6)
-        ### Image
-        for i in range(img0.shape[0]):
-            for j in range(img0.shape[1]):
-                for k in range(img0.shape[2]):
-                    for l in range(img0.shape[3]):
-                        self.assertAlmostEqual(img0[i,j,k,l], img1[i,j,k,l], 6)                
-        ## First image another way
-        hdr1, img1 = db1[0]
-        for attr in ('stokes_params', 'ngrid', 'pixel_size', 'ngrid'):
-            attr0 = getattr(hdr0, attr, None)
-            if isinstance(attr0, bytes):
-                attr0 = attr0.decode()
-            self.assertEqual(attr0, getattr(hdr1, attr, None))
-        for attr in ('start_time', 'int_len', 'lst', 'start_freq', 'stop_freq', 'bandwidth', 'fill', 'center_ra', 'center_dec'):
-            self.assertAlmostEqual(getattr(hdr0, attr, None), getattr(hdr1, attr, None), 6)
-        ### Image
-        for i in range(img0.shape[0]):
-            for j in range(img0.shape[1]):
-                for k in range(img0.shape[2]):
-                    for l in range(img0.shape[3]):
-                        self.assertAlmostEqual(img0[i,j,k,l], img1[i,j,k,l], 6)
-                        
-        db0.close()
-        db1.close()
-        
     def test_bad_freq_list(self):
         """Test the list of bad frequencies that should be flagged/removed."""
